# addons/pilotsuite/app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PilotSuite Core API
  description: |
    REST API for the PilotSuite Core Add-on.
    
    This API provides endpoints for:
    - Habitus pattern mining and rules discovery
    - Brain graph visualization and synchronization
    - User preference learning (MUPL)
    - Tag system v2 for entity classification
    - Mood context aggregation
    - Event ingestion with idempotency
    - System health monitoring
    - Energy monitoring, anomaly detection, and load shifting
    - Zone-based energy tracking and Sankey flow diagrams
    
    ## Authentication
    All endpoints require `X-Auth-Token` header when authentication is enabled.
    
    ## Idempotency
    Event endpoints support `Idempotency-Key` header for deduplication.
  version: 5.4.0
  contact:
    name: PilotSuite
    url: https://github.com/GreenhillEfka/pilotsuite-styx-core
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://homeassistant.local:8909
    description: PilotSuite Core Add-on (default port)
  - url: http://localhost:8909
    description: Development server
...
paths:
...
  /api/v1/habitus/status:
    get:
      tags: [Habitus]
      summary: Get habitus miner status
      operationId: getHabitusStatus
      responses:
        '200':
          description: Habitus miner status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HabitusStatus'
        '500':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/habitus/rules:
    get:
      tags: [Habitus]
      summary: Get discovered A→B rules
      operationId: getHabitusRules
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: min_score
          in: query
          schema:
            type: number
            format: float
        - name: a_filter
          in: query
          schema:
            type: string
        - name: b_filter
          in: query
          schema:
            type: string
        - name: domain_filter
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of discovered rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulesResponse'
  /api/v1/habitus/mine:
    post:
      tags: [Habitus]
      summary: Mine rules from HA events
      operationId: mineRules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MineRequest'
      responses:
        '200':
          description: Mining result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MineResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/mood:
    get:
      tags: [Mood]
      summary: Get current mood context
      operationId: getMood
      parameters:
        - name: zone_id
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Mood context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodResponse'

  # ==================== TAGS ====================
...
  /api/v1/events:
    post:
      tags: [Events]
      summary: Ingest event
      operationId: ingestEvent
      parameters:
        - name: Idempotency-Key
          in: header
          description: Unique key for deduplication
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventIngest'
      responses:
        '200':
          description: Event processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventIngestResponse'

  # ==================== CANDIDATES ====================
...
components:
  schemas:
...
    HabitusStatus:
      type: object
      properties:
        status:
          type: string
          enum: [ok, error]
        version:
          type: string
        statistics:
          type: object
          properties:
            total_rules:
              type: integer
            total_events:
              type: integer
            last_mining_ms:
              type: integer
        config:
          type: object
    Rule:
      type: object
      properties:
        A:
          type: string
          description: Antecedent entity/transition
        B:
          type: string
          description: Consequent entity/transition
        dt_sec:
          type: number
          description: Time delta in seconds
        nA:
          type: integer
          description: Antecedent count
        nB:
          type: integer
          description: Consequent count
        nAB:
          type: integer
          description: Co-occurrence count
        confidence:
          type: number
          minimum: 0
          maximum: 1
        confidence_lb:
          type: number
          description: Lower bound confidence
        lift:
          type: number
        leverage:
          type: number
        score:
          type: number
        evidence:
          type: object
          properties:
            hit_examples:
              type: array
              items:
                type: object
            miss_examples:
              type: array
              items:
                type: object
            latency_quantiles:
              type: array
              items:
                type: number
    RulesResponse:
      type: object
      properties:
        status:
          type: string
        total_rules:
          type: integer
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
    MineRequest:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            type: object
        config:
          type: object
    MineResponse:
      type: object
      properties:
        status:
          type: string
        mining_time_sec:
          type: number
        total_input_events:
          type: integer
        discovered_rules:
          type: integer
        top_rules:
          type: array
          items:
            type: object
...
    MoodResponse:
      type: object
      properties:
        status:
          type: string
        mood:
          type: object
          properties:
            score:
              type: number
              minimum: 0
              maximum: 1
            confidence:
              type: number
            factors:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  value:
                    type: number
                  weight:
                    type: number
        zone_id:
          type: string
        user_id:
          type: string

    # ---------- TAGS ----------
...
    EventIngest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
        text:
          type: string
        payload:
          type: object
        idempotency_key:
          type: string
        timestamp:
          type: string
          format: date-time
    EventIngestResponse:
      type: object
      properties:
        ok:
          type: boolean
        stored:
          type: boolean
        deduped:
          type: boolean
        event_id:
          type: string

    # ---------- CANDIDATES ----------
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        code:
          type: string

    # ---------- NEURONS ----------
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Token
security:
  - ApiKeyAuth: []